streamlit_app/requirements.txt
streamlit
streamlit-audiorec
requests
python-dotenv
fpdf2
pytest
requests-mock
